.github/scripts/requirements.txt
filelock==3.13.1
gitpython==3.1.43
semver==3.0.2
